+++ src/services/Uploader/VideoApi.js
@@ -32,26 +32,4 @@ const askForSubtitles = (videoName) => {
             console.log(error);
         })
 };
-
-const askForSignedImageURL = (pictureName) => {
-    return axios.post(`${URL_API}/picture/upload`, {pictureName: pictureName})
-        .then((result) => {
-            return result.data.signedURL;
-        })
-        .catch((error) => {
-            console.log(error)
-        })
-};
-
-const uploadImageToGCP = (video, signedUrl) => {
-    return axios.put(signedUrl, video, {headers: {'Content-Type': 'image/*'}})
-        .then(() => {
-            return true;
-        })
-        .catch((error) => {
-            console.log(error);
-            return false;
-        })
-};
-
-export {askForSignedURL, uploadVideoToGCP, askForSubtitles, uploadImageToGCP, askForSignedImageURL}
+export {askForSignedURL, uploadVideoToGCP, askForSubtitles}
